Adds input cast validation to `transformer`
Whenever a cast is consumed from `transformer`, a cast validation is now
performed.

This commit makes even more natural to develop the commands as the developer
doesn't need to worry about receiving invalid casts.

Naturally, that's the expectation - there might exist a case when a
a command could potentially *want* to receive an invalid cast (which is
not the situation right now).